requirements.txt
python-telegram-bot==21.7
aiohttp==3.11.11
python-dotenv==1.0.1
x10-python-trading-starknet
